Rename files to rtcMember.ts, rtcMember.test.ts
