+++ packages/react-sdk/src/model/index.ts
@@ -35,22 +35,22 @@ export {
   isValidDocumentSnapshotRoomEvent,
 } from './documentSnapshot';
 export type { DocumentSnapshot } from './documentSnapshot';
+export { STATE_EVENT_ROOM_NAME, isValidRoomNameEvent } from './roomNameEvent';
+export type { RoomNameEvent } from './roomNameEvent';
 export {
   ROOM_EVENT_4143_RTC_MEMBER,
   isLivekitTransport,
   isRtcMemberJoinEvent,
   isRtcMemberLeaveEvent,
   isValidWhiteboardRtcMemberEvent,
-} from './matrixRtcMember';
+} from './rtcMember';
 export type {
   LivekitTransport,
   RtcMember,
   RtcMemberJoin,
   RtcMemberLeave,
   Transport,
-} from './matrixRtcMember';
-export { STATE_EVENT_ROOM_NAME, isValidRoomNameEvent } from './roomNameEvent';
-export type { RoomNameEvent } from './roomNameEvent';
+} from './rtcMember';
 export {
   STATE_EVENT_4143_RTC_SLOT,
   isValidWhiteboardRtcSlotEvent,
