+++ packages/react-sdk/src/model/rtcMember.ts
